+++ routes/api/profile.js
@@ -131,4 +131,20 @@ router.get('/user/:user_id', async (req, res) => {
     }
 });
 
+// @route     DELETE api/profile
+// @desc      Delete profile, user & associated lists
+// @access    Private
+router.delete('/', auth, async (req, res) => {
+    try {
+        await Profile.findOneAndRemove({ user: req.user.id });
+
+        await User.findOneAndRemove({ _id: req.user.id });
+
+        return res.json({ msg: 'User deleted' });
+    } catch (err) {
+        console.error(err.message);
+        return res.status(500).send('Server error');
+    }
+});
+
 module.exports = router;